Rename modules to snake_case
