+++ photon_stream/__init__.py
@@ -1,11 +1,11 @@
-from .EventListReader import EventListReader
-from .SimulationReader import SimulationReader
+from .event_list_reader import EventListReader
+from .simulation_reader import SimulationReader
 
-from .Event import Event
-from .PhotonStream import PhotonStream
-from .ObservationInformation import ObservationInformation
-from .PhotonCluster import PhotonStreamCluster
-from .PhotonCluster import PhotonTimeLineCluster
+from .event import Event
+from .photon_stream import PhotonStream
+from .observation_information import ObservationInformation
+from .photon_cluster import PhotonStreamCluster
+from .photon_cluster import PhotonTimeLineCluster
 
 from .jsonl2binary import jsonl2binary
