+++ photon_stream/geometry.py
