+++ photon_stream/observation_information.py
