+++ photon_stream/photon_stream.py
@@ -1,6 +1,6 @@
 import numpy as np
 from .io import magic_constants
-from .representations import raw_phs_to_point_cloud
+from .representations import raw_phs_to_xyt
 from .representations import raw_phs_to_list_of_lists
 from .representations import raw_phs_to_image_sequence
 from .geometry import GEOMETRY
@@ -18,7 +18,7 @@ class PhotonStream(object):
         return len(self.raw) - magic_constants.NUMBER_OF_PIXELS
 
     @property
-    def point_cloud(self):
+    def xyt(self):
         ''' Returns a Nx3 matrix for N photons in the stream. Each row
         represents a photon in the three dimensinal space of x-direction [rad],
         y-direction [rad], and arrival time [s].
@@ -27,7 +27,7 @@ class PhotonStream(object):
         photon-stream. It is useful for e.g. directly plotting the photon stream
         into its 3 dimensional space, or for density clustering in the stream.
         '''
-        return raw_phs_to_point_cloud(
+        return raw_phs_to_xyt(
             self.raw,
             cx=self.geometry.x_angle,
             cy=self.geometry.y_angle,
@@ -36,14 +36,14 @@ class PhotonStream(object):
     @property
     def list_of_lists(self):
         '''
-        Returns a list along all pixels of lists for each photon arrival time slice. 
+        Returns a list along all pixels of lists for each photon arrival time slice.
         '''
         return raw_phs_to_list_of_lists(self.raw)
 
 
     @property
     def image_sequence(self):
-        return raw_phs_to_image_sequence(self.raw) 
+        return raw_phs_to_image_sequence(self.raw)
 
 
     def __eq__(self, other):
@@ -51,7 +51,7 @@ class PhotonStream(object):
             if not np.abs(self.slice_duration - other.slice_duration) < MAX_RESIDUAL_SLICE_DURATION_NS: return False
 
             # Saturated Pixels
-            if not len(self.saturated_pixels) == len(other.saturated_pixels): return False            
+            if not len(self.saturated_pixels) == len(other.saturated_pixels): return False
             for i, saturated_pixel_in in enumerate(self.saturated_pixels):
                 if not saturated_pixel_in == other.saturated_pixels[i]: return False
