+++ photon_stream/tests/test_PhotonStream.py
@@ -10,8 +10,8 @@ run_path = pkg_resources.resource_filename(
 def test_properties():
     event = next(ps.EventListReader(run_path))
     assert event.photon_stream.number_photons == 4984
-    assert event.photon_stream.point_cloud.shape == (4984, 3)
+    assert event.photon_stream.xyt.shape == (4984, 3)
     assert event.photon_stream.image_sequence.shape == (
-        ps.io.magic_constants.NUMBER_OF_TIME_SLICES, 
+        ps.io.magic_constants.NUMBER_OF_TIME_SLICES,
         ps.io.magic_constants.NUMBER_OF_PIXELS
-    ) 
+    )
